Rename content files. Linting
